requirements.txt
beautifulsoup4==4.9.1
ImageHash==4.1.0
jellyfish==0.8.2
Jinja2==2.11.2
Markdown==3.2.2
markdownify==0.5.2
MarkupSafe==1.1.1
pendulum==2.1.2
Pillow==7.2.0
PyMySQL==0.10.0
python-frontmatter==0.5.0
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
SQLAlchemy==1.3.19
toml==0.10.1
tqdm==4.50.0
typographeur==0.5.0
Wand==0.6.2
